split up "main_utils" into service modules
