+++ src/api/context.py
@@ -1,0 +1,91 @@
+"""Application context dataclasses."""
+
+import pathlib
+from dataclasses import dataclass
+from typing import Optional
+
+import flask
+
+from src.rag.embedding_stores.AbstractEmbeddingStore import ChunkEmbeddingStore
+from src.rag.rag import RAG
+
+
+@dataclass
+class ActiveRAG:
+    rag: RAG
+    vectorstore: ChunkEmbeddingStore
+    process: str
+    ready: bool = False
+    initializing: bool = False
+
+    def switch_readiness(self):
+        self.ready = not self.ready
+
+
+@dataclass
+class ProcessContext:
+    """Runtime process/thread tracking state."""
+
+    running: dict
+    threads: dict
+
+
+@dataclass
+class PipelineContext:
+    """Runtime state for active pipeline objects."""
+
+    active_objects: dict
+
+
+@dataclass
+class StorageContext:
+    """Filesystem locations used by the application."""
+
+    file_storage_dir: pathlib.Path
+
+
+@dataclass
+class RagContext:
+    """Runtime RAG state."""
+
+    active: Optional[ActiveRAG] = None
+
+
+@dataclass
+class AppContext:
+    """Shared Flask application context and grouped runtime state."""
+
+    app: flask.Flask
+    processes: ProcessContext
+    pipeline: PipelineContext
+    storage: StorageContext
+    rag: RagContext
+
+    @property
+    def running_processes(self) -> dict:
+        """Compatibility alias for ``processes.running``."""
+        return self.processes.running
+
+    @property
+    def pipeline_threads_store(self) -> dict:
+        """Compatibility alias for ``processes.threads``."""
+        return self.processes.threads
+
+    @property
+    def current_active_pipeline_objects(self) -> dict:
+        """Compatibility alias for ``pipeline.active_objects``."""
+        return self.pipeline.active_objects
+
+    @property
+    def file_storage_dir(self) -> pathlib.Path:
+        """Compatibility alias for ``storage.file_storage_dir``."""
+        return self.storage.file_storage_dir
+
+    @property
+    def active_rag(self) -> Optional[ActiveRAG]:
+        """Compatibility alias for ``rag.active``."""
+        return self.rag.active
+
+    @active_rag.setter
+    def active_rag(self, value: Optional[ActiveRAG]) -> None:
+        self.rag.active = value
