+++ src/api/responses.py
@@ -1,0 +1,15 @@
+"""HTTP response status constants."""
+
+from enum import IntEnum
+
+
+class HTTPResponses(IntEnum):
+    OK = 200
+    ACCEPTED = 202
+    BAD_REQUEST = 400
+    UNAUTHORIZED = 401
+    FORBIDDEN = 403
+    NOT_FOUND = 404
+    INTERNAL_SERVER_ERROR = 500
+    NOT_IMPLEMENTED = 501
+    SERVICE_UNAVAILABLE = 503
