+++ src/api/routes/pipeline.py
@@ -7,12 +7,12 @@ import pathlib
 from flask import Blueprint, jsonify, request
 
 from src.api.services.configuration import load_configs
-from main_utils import (
-    HTTPResponses,
-    PipelineLanguage,
+from src.api.responses import HTTPResponses
+from src.common.parsing import (
     get_bool_expression,
     string_conformity,
 )
+from src.pipeline.status import PipelineLanguage
 from src.api.pipeline import run_complete_pipeline
