+++ src/pipeline/base.py
@@ -1,194 +1,20 @@
-import logging
+"""Base class for pipeline step utilities."""
+
 import pathlib
-import re
-import threading
 from abc import ABC, abstractmethod
-from collections import namedtuple, defaultdict
 from copy import deepcopy
-from dataclasses import dataclass
-from enum import Enum, IntEnum
 from inspect import getfullargspec
 from pathlib import Path
-from typing import Union, Optional, Any, Callable, Tuple, Iterator
+from typing import Any, Callable, Optional, Tuple, Union
 
 import flask
-import spacy
 import yaml
-from dataclass_wizard import JSONWizard
-from flask import jsonify, Response
+from flask import Response, jsonify
 from munch import Munch, unmunchify
-from waiting import wait, TimeoutExpired
 from werkzeug.datastructures import FileStorage
 
-from src.rag.embedding_stores.AbstractEmbeddingStore import ChunkEmbeddingStore
-from src.rag.rag import RAG
-
-
-class ProcessStatus(str, Enum):
-    STARTED = "started"
-    RUNNING = "running"
-    FINISHED = "finished"
-    ABORTED = "aborted"
-    NOT_PRESENT = "not present"
-    STOPPED = "stopped"
-
-
-class HTTPResponses(IntEnum):
-    OK = 200
-    ACCEPTED = 202
-    BAD_REQUEST = 400
-    UNAUTHORIZED = 401
-    FORBIDDEN = 403
-    NOT_FOUND = 404
-    INTERNAL_SERVER_ERROR = 500
-    NOT_IMPLEMENTED = 501
-    SERVICE_UNAVAILABLE = 503
-
-
-class StepsName:
-    DATA = "data"
-    EMBEDDING = "embedding"
-    CLUSTERING = "clustering"
-    GRAPH = "graph"
-    INTEGRATION = "integration"
-    ALL = [DATA, EMBEDDING, CLUSTERING, GRAPH, INTEGRATION]
-
-
-@dataclass
-class ActiveRAG:
-    rag: RAG
-    vectorstore: ChunkEmbeddingStore
-    process: str
-    ready: bool = False
-    initializing: bool = False
-
-    def switch_readiness(self):
-        self.ready = not self.ready
-
-
-@dataclass
-class ProcessContext:
-    """Runtime process/thread tracking state."""
-
-    running: dict
-    threads: dict
-
-
-@dataclass
-class PipelineContext:
-    """Runtime state for active pipeline objects."""
-
-    active_objects: dict
-
-
-@dataclass
-class StorageContext:
-    """Filesystem locations used by the application."""
-
-    file_storage_dir: pathlib.Path
-
-
-@dataclass
-class RagContext:
-    """Runtime RAG state."""
-
-    active: Optional[ActiveRAG] = None
-
-
-@dataclass
-class AppContext:
-    """Shared Flask application context and grouped runtime state."""
-
-    app: flask.Flask
-    processes: ProcessContext
-    pipeline: PipelineContext
-    storage: StorageContext
-    rag: RagContext
-
-    @property
-    def running_processes(self) -> dict:
-        """Compatibility alias for ``processes.running``."""
-        return self.processes.running
-
-    @property
-    def pipeline_threads_store(self) -> dict:
-        """Compatibility alias for ``processes.threads``."""
-        return self.processes.threads
-
-    @property
-    def current_active_pipeline_objects(self) -> dict:
-        """Compatibility alias for ``pipeline.active_objects``."""
-        return self.pipeline.active_objects
-
-    @property
-    def file_storage_dir(self) -> pathlib.Path:
-        """Compatibility alias for ``storage.file_storage_dir``."""
-        return self.storage.file_storage_dir
-
-    @property
-    def active_rag(self) -> Optional[ActiveRAG]:
-        """Compatibility alias for ``rag.active``."""
-        return self.rag.active
-
-    @active_rag.setter
-    def active_rag(self, value: Optional[ActiveRAG]) -> None:
-        self.rag.active = value
-
-
-# Backwards-compatible alias for code or pickles that still reference the old
-# name. New code should import/use AppContext.
-PersistentObjects = AppContext
-
-
-@dataclass
-class NegspacyConfig(JSONWizard):
-    chunk_prefix: str | list[str] | None = None
-    neg_termset_file: str | None = None
-    scope: int | None = None
-    language: str | None = None
-    feat_of_interest: str | None = None
-
-
-class StoppableThread(threading.Thread):
-    """Thread class with a stop() method. The thread itself has to check
-    regularly for the stopped() condition.
-    From: https://stackoverflow.com/questions/323972/is-there-any-way-to-kill-a-thread
-    """
-
-    def __init__(self, group, target, name, target_args=(), target_kwargs=None):
-        super().__init__(
-            args=target_args,
-            group=group,
-            target=target,
-            name=name,
-            kwargs=target_kwargs,
-        )
-        self._stop_event = threading.Event()
-        self._hard_stop_event = threading.Event()
-        self._return = None
-
-    @property
-    def set_to_stop(self):
-        return self._stop_event.is_set() or self._hard_stop_event.is_set()
-
-    @property
-    def set_to_hard_stop(self):
-        return self._stop_event.is_set() and self._hard_stop_event.is_set()
-
-    @property
-    def return_value(self):
-        if self.is_alive():
-            return False
-        return self._return
-
-    def run(self):
-        if self._target is not None:
-            self._return = self._target(*self._args, **self._kwargs)
-
-    def stop(self, hard_stop=False):
-        self._stop_event.set()
-        if hard_stop:
-            self._hard_stop_event.set()
+from src.common.threads import StoppableThread
+from src.pipeline.status import ProcessStatus, StepsName, add_status_to_running_process
 
 
 class BaseUtil(ABC):
@@ -504,180 +330,3 @@ class BaseUtil(ABC):
                 )
             self._app.logger.error(e)
         return None
-
-
-pipeline_query_params = namedtuple(
-    "PipelineQueryParams",
-    [
-        "process_name",
-        "language",
-        "skip_present",
-        "omitted_pipeline_steps",
-        "return_statistics",
-    ],
-)
-
-steps_relation_dict = {
-    StepsName.DATA: 1,
-    StepsName.EMBEDDING: 2,
-    StepsName.CLUSTERING: 3,
-    StepsName.GRAPH: 4,
-    StepsName.INTEGRATION: 5,
-}
-
-
-class PipelineLanguage:
-    language_map = {
-        "en": "en",
-        "english": "en",
-        "englisch": "en",
-        "de": "de",
-        "german": "de",
-        "deutsch": "de",
-    }
-
-    @staticmethod
-    def language_from_string(lang):
-        return PipelineLanguage.language_map.get(lang.lower(), "en")
-
-
-def add_status_to_running_process(
-    process_name: str,
-    step_name: str,
-    step_status: ProcessStatus,
-    running_processes: dict,
-):
-    _step = {
-        "name": step_name,
-        "rank": steps_relation_dict[step_name],
-        "status": step_status,
-    }
-    _remove = -1
-    if not running_processes.get(process_name, False):
-        running_processes[process_name] = {
-            "name": process_name,
-            "status": [],
-        }
-    else:
-        for _i, _status in enumerate(running_processes[process_name]["status"]):
-            if _status.get("name", False) == step_name:
-                _remove = _i
-                break
-        if _remove >= 0:
-            running_processes[process_name]["status"].pop(_remove)
-
-    if _remove >= 0:
-        running_processes[process_name]["status"].insert(_remove, _step)
-    else:
-        running_processes[process_name]["status"].append(_step)
-    return running_processes
-
-
-def get_bool_expression(str_bool: str, default: Union[bool, str] = False) -> bool:
-    if isinstance(str_bool, bool):
-        return str_bool
-    elif isinstance(str_bool, str):
-        return {
-            "true": True,
-            "yes": True,
-            "y": True,
-            "ja": True,
-            "j": True,
-            "false": False,
-            "no": False,
-            "n": False,
-            "nein": False,
-        }.get(str_bool.lower(), default)
-    else:
-        return False
-
-
-def string_conformity(s: str):
-    if s is None:
-        return None
-    return re.sub(r"\s+|-+", "_", s.lower())
-
-
-def load_spacy_model(spacy_model: str, logger: logging.Logger, default_model: str):
-    def is_valid_spacy_model(model: str):
-        from spacy.cli.download import get_compatibility
-
-        if model in get_compatibility():
-            return True
-        logger.error(f"'{model}' is not a valid model name.")
-        return False
-
-    def wait_for_download(model: str, time_out: int = 30):
-        spacy.cli.download(model)
-        wait_pred = lambda: model in spacy.util.get_installed_models()
-        try:
-            wait(wait_pred, timeout_seconds=time_out)
-        except TimeoutExpired:
-            logger.warning(
-                f"TimeOut while waiting >{time_out} seconds for download to finish."
-                f" Hopefully this is just due to installed models not refreshing."
-            )
-
-    spacy_language = None
-    try:
-        spacy_language = spacy.load(spacy_model)
-    except IOError as e:
-        if spacy_model != default_model:
-            if is_valid_spacy_model(spacy_model):
-                logger.info(
-                    f"Model '{spacy_model}' doesn't seem to be installed; trying to download model."
-                )
-                wait_for_download(spacy_model)
-                spacy_language = spacy.load(spacy_model)
-            else:
-                logger.error(f"{e}\nUsing default model {default_model}.")
-                try:
-                    spacy_language = spacy.load(default_model)
-                except IOError as e:
-                    logger.error(
-                        f"{e}\ntrying to download default model {default_model}."
-                    )
-                    wait_for_download(default_model)
-                    spacy_language = spacy.load(default_model)
-        else:
-            logger.error(f"{e}\ntrying to download default model {default_model}.")
-            wait_for_download(default_model)
-            spacy_language = spacy.load(default_model)
-    return spacy_language
-
-
-def get_default_spacy_model():
-    return "en_core_web_trf"
-
-
-def transform_document_addition_results(iterator: Iterator):
-    phrases_dict = dict()
-    type_list = ["added", "incorporated"]
-    _additional_info_key = "additional_info"
-    _phrases_key = "phrases"
-    _graph_field_key = "graph_cluster"
-    _phrase_id_key = "_id"
-    _text_key = "text"
-    _offsets_key = "offsets"
-
-    for doc_id, graph_dict in iterator:
-        for _type in type_list:
-            for _phrase, _additional_info in zip(
-                graph_dict.get(_type, {}).get(_phrases_key, []),
-                graph_dict.get(_type, {}).get(_additional_info_key, []),
-            ):
-                if _phrase_id := _phrase.get(_phrase_id_key, None):
-                    if _phrase_id not in phrases_dict:
-                        phrases_dict[_phrase_id] = {
-                            "graph": _phrase.get(_graph_field_key, [None])[0],
-                            "id": _phrase_id,
-                            "documents": list(),
-                            "label": _additional_info.get(_text_key, ""),
-                        }
-                    phrases_dict[_phrase_id]["documents"].append(
-                        {
-                            "id": doc_id,
-                            "offsets": _additional_info.get(_offsets_key, []),
-                        }
-                    )
-    return phrases_dict
