+++ src/pipeline/document_addition.py
@@ -11,7 +11,9 @@ import networkx as nx
 import numpy as np
 
 from src.pipeline.load_utils import FactoryLoader
-from main_utils import HTTPResponses, StepsName, transform_document_addition_results
+from src.api.responses import HTTPResponses
+from src.pipeline.document_results import transform_document_addition_results
+from src.pipeline.status import StepsName
 from src.core import data_functions, embedding_functions
 from src.api.request_parsing import document_adding_json
 from src.core.graph_functions import GraphIncorp
