+++ src/pipeline/steps/graph_creation_util.py
@@ -7,7 +7,8 @@ from pyvis import network as net
 from werkzeug.datastructures import FileStorage
 
 from src.pipeline.load_utils import FactoryLoader
-from main_utils import StepsName, BaseUtil
+from src.pipeline.base import BaseUtil
+from src.pipeline.status import StepsName
 from src.core.cluster_functions import WordEmbeddingClustering
 from src.common.util_functions import save_pickle
